Fix #161: Add support for custom elements
This is the first part of upstreaming custom elements into the DOM and
HTML Standards, instead of having them be in a separate spec that is
essentially a bunch of monkeypatches. The second part will be
upstreaming the remaining portions into HTML.

This patch adds an option param into Element's createElement/NS and defines
the behaviors of create/change/remove custom element.

This patch intergrates several patchs in whatwg. Please look into
the bug iusse for more info.