Fix #161: Set the stage for preventing double custom element constructor calls
This is part of w3c/webcomponents#533. It introduces a new custom element
state of "failed", which HTML will use on failed upgrades or failed
parser-initiated element creation, to ensure that if a constructor fails,
the element will not be upgraded in the future (i.e. the constructor will
not be called again for that element).
Also fix a tiny syntax error.